+++ CHANGES.md
@@ -10,6 +10,8 @@ This document is intended for Spotless developers.
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+### Added
+* Added support for Protobuf formatting based on [Buf](https://buf.build/) (#1208).
 
 ## [2.25.3] - 2022-05-10
 ### Fixed